# openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "File Conversion Platform",
    "description": "Comprehensive file conversion platform supporting 100+ conversion types",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/audio/convert/{conversion_id}": {
      "post": {
        "tags": [
          "audio"
        ],
        "summary": "Convert Audio",
        "description": "Convert audio files between formats or perform audio operations.\n\nSupported conversions:\n- MP3 Γåö WAV, AAC, OGG\n- FLAC ΓåÆ MP3\n- M4A ΓåÆ MP3\n- WAV ΓåÆ FLAC\n- OPUS ΓåÆ MP3\n- Text ΓåÆ Audio (TTS)\n- Video ΓåÆ Audio extraction",
        "operationId": "convert_audio_api_audio_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_audio_api_audio_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/code/convert/{conversion_id}": {
      "post": {
        "tags": [
          "code"
        ],
        "summary": "Convert Code",
        "description": "Convert code and data formats.\n\nSupported conversions:\n- JSON Γåö CSV, XML, YAML\n- CSV Γåö XML, JSON\n- Excel ΓåÆ JSON\n- HTML ΓåÆ PDF, Markdown\n- Markdown ΓåÆ HTML\n- Jupyter Notebook ΓåÆ HTML, PDF",
        "operationId": "convert_code_api_code_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_code_api_code_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/design/convert/{conversion_id}": {
      "post": {
        "tags": [
          "design"
        ],
        "summary": "Convert Design",
        "description": "Convert design and CAD files.\n\nSupported conversions:\n- PSD ΓåÆ PNG, JPG\n- SVG ΓåÆ PDF\n- DXF ΓåÆ SVG",
        "operationId": "convert_design_api_design_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_design_api_design_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/convert/{conversion_id}": {
      "post": {
        "tags": [
          "database"
        ],
        "summary": "Convert Database",
        "description": "Convert database and big-data formats.\n\nSupported conversions:\n- SQL Γåö CSV\n- JSON ΓåÆ SQL\n- Excel ΓåÆ SQL\n- Parquet ΓåÆ CSV\n- Avro ΓåÆ JSON",
        "operationId": "convert_database_api_database_convert__conversion_id__post",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_database_api_database_convert__conversion_id__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/download/{conversion_id}": {
      "get": {
        "summary": "Download File",
        "description": "Download converted file",
        "operationId": "download_file_api_download__conversion_id__get",
        "parameters": [
          {
            "name": "conversion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_convert_audio_api_audio_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_audio_api_audio_convert__conversion_id__post"
      },
      "Body_convert_code_api_code_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_code_api_code_convert__conversion_id__post"
      },
      "Body_convert_database_api_database_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_database_api_database_convert__conversion_id__post"
      },
      "Body_convert_design_api_design_convert__conversion_id__post": {
        "properties": {
          "target_format": {
            "type": "string",
            "title": "Target Format"
          }
        },
        "type": "object",
        "required": [
          "target_format"
        ],
        "title": "Body_convert_design_api_design_convert__conversion_id__post"
      },
...
      "ConversionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "target_format": {
            "type": "string",
            "title": "Target Format"
          },
          "file_size": {
            "type": "number",
            "title": "File Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "source_format",
          "target_format",
          "file_size",
          "status",
          "created_at"
        ],
        "title": "ConversionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}